Prefer element reference over method invocation
Ruby presents two ~syntactic~ sugarings that can not be distinguished
unambiguously from syntax alone:

First, array elements can be referenced using a bracketed argument after
any amount of white space, so:

    x.[](0)

    x[0]

    x [0]

are all equivalent.

Second, methods may be invoked with omitted parends, so:

    f(y)

    f y

are equivalent.

The ambiguity can be seen when the function argument is a literal array:

    f [0]

At this point, there is no syntactic information that can distinguish
between element reference and procedural invocation.

This can be seen by running this program in irb:

    irb(main):001:0> x = [0, 1, 2]
    => [0, 1, 2]
    irb(main):002:0> x.[](0)
    => 0
    irb(main):003:0> x [0]
    => 0
    irb(main):004:0> def y(z)
    irb(main):005:1>   z
    irb(main):006:1> end
    => :y
    irb(main):007:0> y([0])
    => [0]
    irb(main):008:0> y [0]
    => [0]

Previously, tree-sitter-ruby handled this ambiguity by presenting both
`x [0]` and `y [0]` as procedural invocation.

However, this causes a parse error as described in
tree-sitter/tree-sitter-ruby#146, when parsing

    d.find { |x| a(x) } [b] == c

Here I add an optional, lower-precedence interpretation of `x [0]` as an
element reference.

Due to the construction of the grammar in this project, this
unfortunately causes problems when attempting to parse constructs like:

    fun [0] do
      something
    end

as the parser will eagerly consume `fun [0]` as the left-hand-side of
a binary expression. To deal with this case, I explicitly add this
construct to the `call` production rule. Unfortunately I had to resort
to the GLR parser in order to resolve the ambiguity between these two
rules.

Finally, note that the tree obtained from the construct

    z [0] == 0

is context-sensitive in Ruby. If `z` is an array type, it is interpreted
as `binary ( reference ( identifier, integer ), integer`. If `z` is
a method, it is interpreted as `call ( identifier, binary ( array
(integer), integer)`. Since tree-sitter assumes the parsed language is
context-free, there's no good way for us to resolve this ambiguity. This
commit prefers the second, method-invocation, interpretation, which
appears to be more common within the test corpus.