Use external scanner logic to distinguish between arrays & subscripts
When an opening square bracket appears immediately after a callable
expression like "a" or "a.b", we must decide between two possible
interpretations of the bracket:
1. It could be part of an element reference, as in
   `a[0] = true`.
2. Or it could be an array literal, passed as an argumet, as in
   `puts [1, 2, 3]`

If there is no preceding whitespace, the bracket should *always* be
treated as part of an element reference. This matches MRI's behavior.

If there *is* preceding whitespace, MRI makes its decision in a
context-sensitive way, based on whether the preceding expression
is a local variable or a method name.

This parser is not context-sensitive, so we instead will interpret
the bracket as part of an array literal whenever that is syntactically
valid, and interpret it as part of element reference otherwise. The
external scanner can use the validity of other expression tokens like
`string` to infer whether an array literal would be valid.